+++ src-tauri/src/binding/function/query/execute.rs
@@ -1,23 +1,15 @@
 use serde::Serialize;
-use serde_json::Value as JsonValue;
 use std::collections::HashMap;
-use uuid::Uuid;
 
 use crate::{
-    Database, LogLevel, auth::{connection::load_connections, token::get_access_token}, binding::model::{
-        dataverse::{
-            entity::{Entity, ResultRow, Value as RowValue},
-            entityattribute::EntityAttribute,
-            entitydefinition::EntityDefinition,
-        },
+    auth::{connection::load_connections, token::get_access_token},
+    binding::model::{
+        dataverse::entity::{Entity, ResultRow, Value as RowValue},
         executesqlrequest::ExecuteSqlRequest,
         executesqlresponse::{ExecuteSqlResponse, SqlQueryMetadata},
-        deletesqlexecuteresponse::DeleteSqlExecuteResponse,
-        deletesqlpreviewresponse::DeleteSqlPreviewResponse,
-        updatesqlexecuteresponse::UpdateSqlExecuteResponse,
-        updatesqlpreviewresponse::UpdateSqlPreviewResponse,
-        response::MultipleResponse,
-    }, dataverse::queryengine::QueryEngine, sql
+    },
+    dataverse::queryengine::QueryEngine,
+    sql, Database, LogLevel,
 };
 
 #[derive(Debug, Serialize)]
@@ -153,555 +145,6 @@ pub async fn execute_sql(
     })
 }
 
-#[tauri::command]
-pub async fn prepare_update_sql(
-    _window: tauri::Window,
-    sql: String,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<UpdateSqlPreviewResponse, String> {
-    if matches!(context.log_level, LogLevel::Debug) {
-        println!("SQL: {}", sql);
-    }
-
-    let stmt = sql::parse_update(&sql).map_err(|e| e.to_string())?;
-
-    if stmt.filter.is_none() {
-        return Err("UPDATE statements must include a WHERE clause.".to_string());
-    }
-
-    if stmt.assignments.is_empty() {
-        return Err("UPDATE statements must include at least one SET assignment.".to_string());
-    }
-
-    let connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-
-    let (entity_set, entity_logical) = sql::resolve_entity_names(&stmt.entity);
-    let primary_id_attribute =
-        resolve_primary_id_attribute(&query_engine, &entity_logical, &entity_set).await?;
-
-    if stmt
-        .assignments
-        .iter()
-        .any(|assignment| normalize_ident(&assignment.column) == normalize_ident(&primary_id_attribute))
-    {
-        return Err("UPDATE cannot modify the primary ID attribute.".to_string());
-    }
-
-    let fetch = sql::update_to_fetchxml(&stmt, &primary_id_attribute)
-        .map_err(|e| e.to_string())?;
-
-    let entities = query_engine
-        .retrieve_multiple_fetchxml(&fetch.entity_set, &fetch.fetchxml)
-        .await
-        .map_err(|error| {
-            println!("Error: {error}");
-            error
-        })?;
-
-    let mut ids: Vec<String> = Vec::new();
-    for entity in entities.value {
-        let value = entity
-            .attributes
-            .get(&primary_id_attribute)
-            .ok_or_else(|| "Primary ID attribute missing from results".to_string())?;
-        let id = value_to_string(value)
-            .ok_or_else(|| "Primary ID attribute was null".to_string())?;
-        ids.push(id);
-    }
-
-    let updates = build_update_attributes(
-        &stmt.assignments,
-        &stmt.entity,
-        stmt.entity_alias.as_deref(),
-    )?;
-
-    let token = Uuid::new_v4().to_string();
-    let batch = crate::UpdateBatch {
-        connection_id,
-        entity_set,
-        entity_logical,
-        primary_id_attribute,
-        ids,
-        updates,
-    };
-    let count = batch.ids.len();
-
-    {
-        let mut batches = database
-            .update_batches
-            .lock()
-            .map_err(|_| "Failed to lock update state".to_string())?;
-        batches.insert(token.clone(), batch);
-    }
-
-    Ok(UpdateSqlPreviewResponse {
-        success: true,
-        message: "Update preview ready.".to_string(),
-        count,
-        token,
-    })
-}
-
-#[tauri::command]
-pub async fn execute_update_sql(
-    _window: tauri::Window,
-    token: String,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<UpdateSqlExecuteResponse, String> {
-    let batch = {
-        let mut batches = database
-            .update_batches
-            .lock()
-            .map_err(|_| "Failed to lock update state".to_string())?;
-        batches
-            .remove(&token)
-            .ok_or_else(|| "Update batch not found or expired.".to_string())?
-    };
-
-    let current_connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    if current_connection_id != batch.connection_id {
-        return Err("Connection changed since preview; re-run the update preview.".to_string());
-    }
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&current_connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-
-    let mut updated = 0usize;
-    let mut failed = 0usize;
-    let mut errors: Vec<String> = Vec::new();
-
-    for id in &batch.ids {
-        let result = query_engine
-            .update_entity(&batch.entity_set, id, &batch.updates)
-            .await;
-        match result {
-            Ok(_) => updated += 1,
-            Err(error) => {
-                failed += 1;
-                errors.push(error);
-            }
-        }
-    }
-
-    let success = failed == 0;
-    let message = if success {
-        format!("Updated {} record(s).", updated)
-    } else {
-        format!("Updated {} record(s) with {} error(s).", updated, failed)
-    };
-
-    Ok(UpdateSqlExecuteResponse {
-        success,
-        message,
-        updated,
-        failed,
-        errors,
-    })
-}
-
-#[tauri::command]
-pub async fn discard_update_sql(
-    _window: tauri::Window,
-    token: String,
-    database: tauri::State<'_, Database>,
-) -> Result<bool, String> {
-    let mut batches = database
-        .update_batches
-        .lock()
-        .map_err(|_| "Failed to lock update state".to_string())?;
-    Ok(batches.remove(&token).is_some())
-}
-
-#[tauri::command]
-pub async fn prepare_delete_sql(
-    _window: tauri::Window,
-    sql: String,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<DeleteSqlPreviewResponse, String> {
-    if matches!(context.log_level, LogLevel::Debug) {
-        println!("SQL: {}", sql);
-    }
-
-    let stmt = sql::parse_delete(&sql).map_err(|e| e.to_string())?;
-
-    if stmt.filter.is_none() {
-        return Err("DELETE statements must include a WHERE clause.".to_string());
-    }
-
-    let connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-
-    let (entity_set, entity_logical) = sql::resolve_entity_names(&stmt.entity);
-    let primary_id_attribute =
-        resolve_primary_id_attribute(&query_engine, &entity_logical, &entity_set).await?;
-
-    let fetch = sql::delete_to_fetchxml(&stmt, &primary_id_attribute)
-        .map_err(|e| e.to_string())?;
-
-    let entities = query_engine
-        .retrieve_multiple_fetchxml(&fetch.entity_set, &fetch.fetchxml)
-        .await
-        .map_err(|error| {
-            println!("Error: {error}");
-            error
-        })?;
-
-    let mut ids: Vec<String> = Vec::new();
-    for entity in entities.value {
-        let value = entity
-            .attributes
-            .get(&primary_id_attribute)
-            .ok_or_else(|| "Primary ID attribute missing from results".to_string())?;
-        let id = value_to_string(value)
-            .ok_or_else(|| "Primary ID attribute was null".to_string())?;
-        ids.push(id);
-    }
-
-    let token = Uuid::new_v4().to_string();
-    let batch = crate::DeleteBatch {
-        connection_id,
-        entity_set,
-        entity_logical,
-        primary_id_attribute,
-        ids,
-    };
-    let count = batch.ids.len();
-
-    {
-        let mut batches = database
-            .delete_batches
-            .lock()
-            .map_err(|_| "Failed to lock delete state".to_string())?;
-        batches.insert(token.clone(), batch);
-    }
-
-    Ok(DeleteSqlPreviewResponse {
-        success: true,
-        message: "Delete preview ready.".to_string(),
-        count,
-        token,
-    })
-}
-
-#[tauri::command]
-pub async fn execute_delete_sql(
-    _window: tauri::Window,
-    token: String,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<DeleteSqlExecuteResponse, String> {
-    let batch = {
-        let mut batches = database
-            .delete_batches
-            .lock()
-            .map_err(|_| "Failed to lock delete state".to_string())?;
-        batches
-            .remove(&token)
-            .ok_or_else(|| "Delete batch not found or expired.".to_string())?
-    };
-
-    let current_connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    if current_connection_id != batch.connection_id {
-        return Err("Connection changed since preview; re-run the delete preview.".to_string());
-    }
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&current_connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-
-    let mut deleted = 0usize;
-    let mut failed = 0usize;
-    let mut errors: Vec<String> = Vec::new();
-
-    for id in &batch.ids {
-        let result = query_engine
-            .delete_entity(&batch.entity_set, id)
-            .await;
-        match result {
-            Ok(_) => deleted += 1,
-            Err(error) => {
-                failed += 1;
-                errors.push(error);
-            }
-        }
-    }
-
-    let success = failed == 0;
-    let message = if success {
-        format!("Deleted {} record(s).", deleted)
-    } else {
-        format!("Deleted {} record(s) with {} error(s).", deleted, failed)
-    };
-
-    Ok(DeleteSqlExecuteResponse {
-        success,
-        message,
-        deleted,
-        failed,
-        errors,
-    })
-}
-
-#[tauri::command]
-pub async fn discard_delete_sql(
-    _window: tauri::Window,
-    token: String,
-    database: tauri::State<'_, Database>,
-) -> Result<bool, String> {
-    let mut batches = database
-        .delete_batches
-        .lock()
-        .map_err(|_| "Failed to lock delete state".to_string())?;
-    Ok(batches.remove(&token).is_some())
-}
-
-#[tauri::command]
-pub async fn list_entity_definitions(
-    _window: tauri::Window,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<MultipleResponse<EntityDefinition>, String> {
-    let connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-    let value = query_engine.list_entity_definitions().await?;
-
-    Ok(MultipleResponse {
-        message: "Metadata retrieved.".to_string(),
-        success: true,
-        value,
-    })
-}
-
-#[tauri::command]
-pub async fn list_entity_attributes(
-    _window: tauri::Window,
-    logical_name: String,
-    database: tauri::State<'_, Database>,
-    context: tauri::State<'_, crate::LaunchContext>,
-) -> Result<MultipleResponse<EntityAttribute>, String> {
-    let connection_id = {
-        let selected = database
-            .selected_connection_id
-            .lock()
-            .map_err(|_| "Failed to lock connection state".to_string())?;
-        selected.ok_or("No connection selected")?
-    };
-
-    let connections = load_connections()?;
-    let connection = connections
-        .into_iter()
-        .find(|connection| connection.id().as_ref() == Some(&connection_id))
-        .ok_or("Connection not found")?;
-
-    let token = get_access_token(&connection, &database).await?;
-
-    let dataverse_url = connection.dataverse_url();
-
-    if dataverse_url.trim().is_empty() {
-        return Err("Connection is missing a Dataverse URL".to_string());
-    }
-
-    let query_engine = QueryEngine::new(&dataverse_url, &token, context.log_level);
-    let value = query_engine
-        .list_entity_attributes(&logical_name)
-        .await?;
-
-    Ok(MultipleResponse {
-        message: "Attributes retrieved.".to_string(),
-        success: true,
-        value,
-    })
-}
-
-async fn resolve_primary_id_attribute(
-    query_engine: &QueryEngine,
-    entity_logical: &str,
-    entity_set: &str,
-) -> Result<String, String> {
-    let definitions = query_engine.list_entity_definitions().await?;
-    let target_logical = normalize_ident(entity_logical);
-    let target_set = normalize_ident(entity_set);
-
-    let matched = definitions.iter().find(|definition| {
-        normalize_ident(&definition.logical_name) == target_logical
-            || normalize_ident(&definition.schema_name) == target_logical
-            || normalize_ident(&definition.entity_set_name) == target_logical
-            || normalize_ident(&definition.entity_set_name) == target_set
-    });
-
-    Ok(matched
-        .and_then(|definition| definition.primary_id_attribute.clone())
-        .unwrap_or_else(|| format!("{}id", entity_logical)))
-}
-
-fn build_update_attributes(
-    assignments: &[sql::UpdateAssignment],
-    entity: &str,
-    entity_alias: Option<&str>,
-) -> Result<HashMap<String, JsonValue>, String> {
-    let mut updates: HashMap<String, JsonValue> = HashMap::new();
-    for assignment in assignments {
-        let column = normalize_update_column(&assignment.column, entity, entity_alias);
-        let value = literal_to_json(&assignment.value)?;
-        updates.insert(column, value);
-    }
-    Ok(updates)
-}
-
-fn normalize_update_column(raw: &str, entity: &str, entity_alias: Option<&str>) -> String {
-    if let Some((table, column)) = split_qualified(raw) {
-        if let Some(table) = table {
-            if table.eq_ignore_ascii_case(entity) {
-                return column.to_string();
-            }
-            if let Some(alias) = entity_alias {
-                if table.eq_ignore_ascii_case(alias) {
-                    return column.to_string();
-                }
-            }
-        }
-    }
-    raw.to_string()
-}
-
-fn split_qualified(value: &str) -> Option<(Option<&str>, &str)> {
-    let mut parts = value.split('.');
-    let first = parts.next()?;
-    let second = parts.next();
-    if let Some(second) = second {
-        Some((Some(first), second))
-    } else {
-        Some((None, first))
-    }
-}
-
-fn literal_to_json(literal: &sql::Literal) -> Result<JsonValue, String> {
-    match literal {
-        sql::Literal::String(value) => Ok(JsonValue::String(value.clone())),
-        sql::Literal::Number(value) => Ok(JsonValue::Number((*value).into())),
-        sql::Literal::Boolean(value) => Ok(JsonValue::Bool(*value)),
-        sql::Literal::Null => Ok(JsonValue::Null),
-    }
-}
-
-fn value_to_string(value: &RowValue) -> Option<String> {
-    match value {
-        RowValue::String(value) => Some(value.clone()),
-        RowValue::Int(value) => Some(value.to_string()),
-        RowValue::Float(value) => Some(value.to_string()),
-        RowValue::Boolean(value) => Some(value.to_string()),
-        RowValue::Null => None,
-    }
-}
-
-fn normalize_ident(value: &str) -> String {
-    value
-        .trim_matches(|ch| ch == '[' || ch == ']' || ch == '"' || ch == '`')
-        .to_ascii_lowercase()
-}
-
 fn entity_to_result_row(entity: Entity, columns_order: &[String]) -> ResultRow {
     let mut attributes = std::collections::HashMap::new();
     for (key, value) in entity.attributes {
@@ -719,9 +162,7 @@ fn entity_to_result_row(entity: Entity, columns_order: &[String]) -> ResultRow {
 
     ensure_columns(&mut attributes, columns_order);
 
-    ResultRow {
-        attributes,
-    }
+    ResultRow { attributes }
 }
 
 const ROW_NUMBER_ATTRIBUTE: &str = "__rownum";
@@ -829,7 +270,8 @@ fn aggregate_fallback_plan(stmt: &sql::SelectStmt) -> Option<AggregatePlan> {
         _ => return None,
     };
 
-    let has_aggregate = items.iter().any(|item| matches!(item.kind, sql::SelectItemKind::Aggregate(_)));
+    let has_aggregate =
+        items.iter().any(|item| matches!(item.kind, sql::SelectItemKind::Aggregate(_)));
     if !has_aggregate {
         return None;
     }
@@ -1141,8 +583,12 @@ fn numeric_value(value: &RowValue) -> Option<(f64, bool)> {
 fn compare_values(left: &RowValue, right: &RowValue) -> std::cmp::Ordering {
     match (left, right) {
         (RowValue::Int(a), RowValue::Int(b)) => a.cmp(b),
-        (RowValue::Int(a), RowValue::Float(b)) => (*a as f64).partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal),
-        (RowValue::Float(a), RowValue::Int(b)) => a.partial_cmp(&(*b as f64)).unwrap_or(std::cmp::Ordering::Equal),
+        (RowValue::Int(a), RowValue::Float(b)) => {
+            (*a as f64).partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal)
+        }
+        (RowValue::Float(a), RowValue::Int(b)) => {
+            a.partial_cmp(&(*b as f64)).unwrap_or(std::cmp::Ordering::Equal)
+        }
         (RowValue::Float(a), RowValue::Float(b)) => a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal),
         (RowValue::String(a), RowValue::String(b)) => a.cmp(b),
         (RowValue::Boolean(a), RowValue::Boolean(b)) => a.cmp(b),
