+++ src-tauri/src/binding/function/query/mod.rs
@@ -1,0 +1,10 @@
+mod delete;
+mod execute;
+mod helpers;
+mod metadata;
+mod update;
+
+pub use delete::{discard_delete_sql, execute_delete_sql, prepare_delete_sql};
+pub use execute::{execute_sql, parse_sql_to_fetchxml, FetchXmlPreview};
+pub use metadata::{list_entity_attributes, list_entity_definitions};
+pub use update::{discard_update_sql, execute_update_sql, prepare_update_sql};
